requirements.txt
python-multipart
fastapi
pydantic
psycopg2-binary